requirements.txt
#
# Please do not manually update this file since the requirements are managed
# by pipenv through Pipfile and Pipfile.lock .
#
# This file is created and managed automatically by the template and it is left
# here only for backwards compatibility reasons with python's ecosystem.
#
# Please use Pipfile to update the requirements.
#
requests>=2.28.2
pycryptodome>=3.17
python-dateutil>=2.8.2
backoff>=1.11.1